chore: rename web package to controller
Rename web/ package to controller/ for better comprehension.
