+++ src/main/java/practice/recipebase/controller/LoginController.java
@@ -1,4 +1,4 @@
-package practice.recipebase.web;
+package practice.recipebase.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
